docs(boundary): generalize private consumer boundary
